+++ packages/video_player/video_player_android/android/src/main/kotlin/io/flutter/plugins/videoplayer/Messages.kt
@@ -300,7 +300,8 @@ data class CreationOptions(
     val uri: String,
     val formatHint: PlatformVideoFormat? = null,
     val httpHeaders: Map<String, String>,
-    val userAgent: String? = null
+    val userAgent: String? = null,
+    val forwardBufferDurationMs: Long? = null
 ) {
   companion object {
     fun fromList(pigeonVar_list: List<Any?>): CreationOptions {
@@ -308,7 +309,8 @@ data class CreationOptions(
       val formatHint = pigeonVar_list[1] as PlatformVideoFormat?
       val httpHeaders = pigeonVar_list[2] as Map<String, String>
       val userAgent = pigeonVar_list[3] as String?
-      return CreationOptions(uri, formatHint, httpHeaders, userAgent)
+      val forwardBufferDurationMs = pigeonVar_list[4] as Long?
+      return CreationOptions(uri, formatHint, httpHeaders, userAgent, forwardBufferDurationMs)
     }
   }
 
@@ -318,6 +320,7 @@ data class CreationOptions(
         formatHint,
         httpHeaders,
         userAgent,
+        forwardBufferDurationMs,
     )
   }