+++ packages/video_player/video_player_android/lib/src/android_video_player.dart
@@ -103,6 +103,7 @@ class AndroidVideoPlayer extends VideoPlayerPlatform {
       httpHeaders: httpHeaders,
       userAgent: userAgent,
       formatHint: formatHint,
+      forwardBufferDurationMs: options.forwardBufferDuration?.inMilliseconds,
     );
 
     final int playerId;