+++ packages/video_player/video_player_avfoundation/lib/src/video_player_plugin_messages.g.dart
@@ -135,14 +135,16 @@ class PlatformVideoViewCreationParams {
 }
 
 class CreationOptions {
-  CreationOptions({required this.uri, required this.httpHeaders});
+  CreationOptions({required this.uri, required this.httpHeaders, this.forwardBufferDurationMs});
 
   String uri;
 
   Map<String, String> httpHeaders;
 
+  int? forwardBufferDurationMs;
+
   List<Object?> _toList() {
-    return <Object?>[uri, httpHeaders];
+    return <Object?>[uri, httpHeaders, forwardBufferDurationMs];
   }
 
   Object encode() {
@@ -154,6 +156,7 @@ class CreationOptions {
     return CreationOptions(
       uri: result[0]! as String,
       httpHeaders: (result[1]! as Map<Object?, Object?>).cast<String, String>(),
+      forwardBufferDurationMs: result[2] as int?,
     );
   }